requirements.txt
glueops-helpers @ https://github.com/GlueOps/python-glueops-helpers-library/archive/refs/tags/v0.1.0.zip
fastapi==0.103.2
uvicorn[standard]==0.23.2
gunicorn==21.2.0
deepdiff==6.6.0
